Move TaskExportData task into a separate file export-data.go
