+++ task/export.go
@@ -90,31 +90,6 @@ func TaskExport(tctx *TaskContext) (*TaskHandlerOutput, error) {
 	}, nil
 }
 
-func TaskExportData(tctx *TaskContext) (*TaskHandlerOutput, error) {
-	var (
-		params       = tctx.Params.ExportData
-		customParams = params.Custom
-		ipfsParams   = params.IPFS
-	)
-	content := strings.NewReader(params.Content)
-	name := fmt.Sprintf("%s-%s", params.Type, params.ID)
-	contentType := "application/json"
-	cid, _, err := uploadFile(tctx, customParams, ipfsParams, name, content, contentType)
-	if err != nil {
-		return nil, err
-	}
-
-	return &TaskHandlerOutput{
-		TaskOutput: &data.TaskOutput{
-			ExportData: &data.ExportDataTaskOutput{
-				IPFS: &data.IPFSExportDataInfo{
-					CID: cid,
-				},
-			},
-		},
-	}, nil
-}
-
 func uploadNftMetadata(tctx *TaskContext, ipfsParams *api.ExportIPFSParams, cid string) (string, error) {
 	var (
 		params = tctx.Task.Params.Export
